Code optimization + UX changes
- Cleaned up codebase by breaking up large classes
- Simplified the permissions request onboarding process
- Added visual feedback during permissions request
