+++ app/src/main/res/layout/dialog_permission_secure_settings.xml
@@ -61,4 +61,12 @@
 		</LinearLayout>
 	</com.google.android.material.card.MaterialCardView>
 
+	<com.google.android.material.loadingindicator.LoadingIndicator
+		android:id="@+id/progressElevatedGrant"
+		android:layout_width="wrap_content"
+		android:layout_height="wrap_content"
+		android:layout_gravity="center"
+		android:layout_marginTop="24dp"
+		android:visibility="gone" />
+
 </LinearLayout>
